+++ player/js/elements/ShapeElement.js
@@ -257,7 +257,7 @@ IShapeElement.prototype.renderModifiers = function() {
     len = this.shapeModifiers.length;
 
     for(i=len-1;i>=0;i-=1){
-        this.shapeModifiers[i].processShapes();
+        this.shapeModifiers[i].processShapes(this.firstFrame);
     }
 };