+++ player/js/renderers/CanvasRenderer.js
@@ -187,20 +187,38 @@ CanvasRenderer.prototype.updateContainerSize = function () {
         elementWidth = this.canvasContext.canvas.width * this.renderConfig.dpr;
         elementHeight = this.canvasContext.canvas.height * this.renderConfig.dpr;
     }
-    if(this.renderConfig.preserveAspectRatio == 'xMidYMid meet'){
-        var elementRel = elementWidth/elementHeight;
-        var animationRel = this.transformCanvas.w/this.transformCanvas.h;
-        if(animationRel>elementRel){
+    var elementRel,animationRel;
+    if(this.renderConfig.preserveAspectRatio.indexOf('meet') !== -1 || this.renderConfig.preserveAspectRatio.indexOf('slice') !== -1){
+        var par = this.renderConfig.preserveAspectRatio.split(' ');
+        var fillType = par[1] || 'meet';
+        var pos = par[0] || 'xMidYMid';
+        var xPos = pos.substr(0,4);
+        var yPos = pos.substr(4);
+        elementRel = elementWidth/elementHeight;
+        animationRel = this.transformCanvas.w/this.transformCanvas.h;
+        if(animationRel>elementRel && fillType === 'meet' || animationRel<elementRel && fillType === 'slice'){
             this.transformCanvas.sx = elementWidth/(this.transformCanvas.w/this.renderConfig.dpr);
             this.transformCanvas.sy = elementWidth/(this.transformCanvas.w/this.renderConfig.dpr);
-            this.transformCanvas.tx = 0;
-            this.transformCanvas.ty = ((elementHeight-this.transformCanvas.h*(elementWidth/this.transformCanvas.w))/2)*this.renderConfig.dpr;
         }else{
             this.transformCanvas.sx = elementHeight/(this.transformCanvas.h / this.renderConfig.dpr);
             this.transformCanvas.sy = elementHeight/(this.transformCanvas.h / this.renderConfig.dpr);
+        }
+
+        if(xPos === 'xMid' && ((animationRel<elementRel && fillType==='meet') || (animationRel>elementRel && fillType === 'slice'))){
             this.transformCanvas.tx = (elementWidth-this.transformCanvas.w*(elementHeight/this.transformCanvas.h))/2*this.renderConfig.dpr;
+        } else if(xPos === 'xMax' && ((animationRel<elementRel && fillType==='meet') || (animationRel>elementRel && fillType === 'slice'))){
+            this.transformCanvas.tx = (elementWidth-this.transformCanvas.w*(elementHeight/this.transformCanvas.h))*this.renderConfig.dpr;
+        } else {
+            this.transformCanvas.tx = 0;
+        }
+        if(yPos === 'YMid' && ((animationRel>elementRel && fillType==='meet') || (animationRel<elementRel && fillType === 'slice'))){
+            this.transformCanvas.ty = ((elementHeight-this.transformCanvas.h*(elementWidth/this.transformCanvas.w))/2)*this.renderConfig.dpr;
+        } else if(yPos === 'YMax' && ((animationRel>elementRel && fillType==='meet') || (animationRel<elementRel && fillType === 'slice'))){
+            this.transformCanvas.ty = ((elementHeight-this.transformCanvas.h*(elementWidth/this.transformCanvas.w)))*this.renderConfig.dpr;
+        } else {
             this.transformCanvas.ty = 0;
         }
+
     }else if(this.renderConfig.preserveAspectRatio == 'none'){
         this.transformCanvas.sx = elementWidth/(this.transformCanvas.w/this.renderConfig.dpr);
         this.transformCanvas.sy = elementHeight/(this.transformCanvas.h/this.renderConfig.dpr);